requirements.txt
python-dotenv
colorama
selenium